refactor: plugins organization
